swizzle UIScene property implementation
