+++ wire-ios/Wire-iOS Tests/NetworkStatusView/NetworkStatusViewSnapshotTests.swift
@@ -1,0 +1,60 @@
+//
+// Wire
+// Copyright (C) 2024 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import XCTest
+
+@testable import Wire
+
+final class NetworkStatusViewSnapShotTests: XCTestCase {
+
+    var sut: NetworkStatusView!
+    var mockContainer: MockContainer!
+
+    override func setUp() {
+        super.setUp()
+
+        accentColor = .purple
+        mockContainer = MockContainer()
+        sut = NetworkStatusView()
+        sut.overrideUserInterfaceStyle = .light
+        sut.backgroundColor = .clear
+        sut.delegate = mockContainer
+    }
+
+    override func tearDown() {
+        sut = nil
+        mockContainer = nil
+
+        super.tearDown()
+    }
+
+    func testOfflineExpandedState() {
+        // GIVEN
+        sut.state = .offlineExpanded
+        // WHEN && THEN
+        verifyInAllPhoneWidths(matching: sut)
+    }
+
+    func testOnlineSynchronizing() {
+        // GIVEN
+        sut.state = .onlineSynchronizing
+        sut.layer.speed = 0 // freeze animations for deterministic tests
+        // WHEN && THEN
+        verifyInAllPhoneWidths(matching: sut)
+    }
+}
