+++ wire-ios/Wire-iOS Tests/NetworkStatusView/NetworkStatusViewTests.swift
@@ -17,7 +17,6 @@
 //
 
 import XCTest
-import WireSystemSupport
 
 @testable import Wire
 
@@ -34,36 +33,38 @@ class MockContainer: NetworkStatusViewDelegate {
 final class NetworkStatusViewTests: XCTestCase {
 
     private var sut: NetworkStatusView!
-    private var mockSceneActivationStateProvider: MockSceneActivationStateProviding!
     private var mockContainer: MockContainer!
 
     override func setUp() {
         super.setUp()
 
-        mockSceneActivationStateProvider = .init()
-        mockSceneActivationStateProvider.activationStateForSceneOf_MockValue = .foregroundActive
-
         mockContainer = MockContainer()
 
-        sut = NetworkStatusView(sceneActivationStateProvider: mockSceneActivationStateProvider)
+        sut = .init()
         sut.delegate = mockContainer
     }
 
     override func tearDown() {
         sut = nil
-        mockSceneActivationStateProvider = nil
         mockContainer = nil
 
         super.tearDown()
     }
 
-    func testThatSyncBarChangesToHiddenWhenTheAppGoesToBackground() {
+    func testThatSyncBarChangesToHiddenWhenTheAppGoesToBackground() throws {
         // GIVEN
         sut.state = .onlineSynchronizing
         XCTAssertEqual(sut.connectingView.heightConstraint.constant, CGFloat.SyncBar.height, "NetworkStatusView should not be zero height")
 
+        // ... the activation state of the scene returns `.background`
+        let getterSelector = #selector(getter: UIScene.activationState)
+        let getBackgroundSelector = #selector(getter: UIScene.backgroundActivationState)
+        let originalGetter = try XCTUnwrap(class_getInstanceMethod(UIScene.self, getterSelector))
+        let temporaryGetter = try XCTUnwrap(class_getInstanceMethod(UIScene.self, getBackgroundSelector))
+        method_exchangeImplementations(originalGetter, temporaryGetter)
+        defer { method_exchangeImplementations(originalGetter, temporaryGetter) }
+
         // WHEN
-        mockSceneActivationStateProvider.activationStateForSceneOf_MockValue = .background
         sut.state = .onlineSynchronizing
 
         // THEN
@@ -71,40 +72,11 @@ final class NetworkStatusViewTests: XCTestCase {
     }
 }
 
-final class NetworkStatusViewSnapShotTests: XCTestCase {
+// MARK: - Method Swizzling
 
-    var sut: NetworkStatusView!
-    var mockContainer: MockContainer!
+private extension UIScene {
 
-    override func setUp() {
-        super.setUp()
-        accentColor = .purple
-        mockContainer = MockContainer()
-        sut = NetworkStatusView()
-        sut.overrideUserInterfaceStyle = .light
-        sut.backgroundColor = .clear
-        sut.delegate = mockContainer
+    @objc var backgroundActivationState: UIScene.ActivationState {
+        .background
     }
-
-    override func tearDown() {
-        sut = nil
-        mockContainer = nil
-        super.tearDown()
-    }
-
-    func testOfflineExpandedState() {
-        // GIVEN
-        sut.state = .offlineExpanded
-        // WHEN && THEN
-        verifyInAllPhoneWidths(matching: sut)
-    }
-
-    func testOnlineSynchronizing() {
-        // GIVEN
-        sut.state = .onlineSynchronizing
-        sut.layer.speed = 0 // freeze animations for deterministic tests
-        // WHEN && THEN
-        verifyInAllPhoneWidths(matching: sut)
-    }
-
 }
